refactor(ui): Move the infinite list helpers out of the hook file
`use-infinite-list.ts` holds three things that are not hooks: the
`PagedResponse` and `InfiniteList` types and the `getNextPageParam`
function. The next change needs a helper that combines several infinite
lists into one, which is a plain function as well and has nowhere to go
in a file of hooks.

Move the three to a new `ui/src/lib/infinite-list.ts` and import them
back, so that the hook file exports the hook alone, which is the usual
React convention. `PagedResponse` is exported now, as the hook reads it
from another file.

`useTotalCount` in `siblings.tsx` describes a page of a paged endpoint
with an inline type of its own, which is what `PagedResponse` is,
so take it into use there and drop the import from `@/api/types.gen`.
