+++ ui/src/hooks/use-infinite-list.ts
@@ -25,16 +25,11 @@ import {
   type SkipToken,
 } from '@tanstack/react-query';
 
-import { PagingData } from '@/api/types.gen';
-
-/**
- * The shape shared by all the paged endpoints of the API, that is, by every generated
- * `PagedResponse*` type.
- */
-type PagedResponse<TItem> = {
-  data: Array<TItem>;
-  pagination: PagingData;
-};
+import {
+  getNextPageParam,
+  type InfiniteList,
+  type PagedResponse,
+} from '@/lib/infinite-list';
 
 /**
  * The parts of the generated `*InfiniteOptions` the hook actually needs. Everything else they carry
@@ -59,32 +54,6 @@ type InfiniteListQueryOptions = {
   staleTime?: number;
   gcTime?: number;
 };
-
-/**
- * What the hook gives back, so that components can take a list without repeating the generics.
- */
-export type InfiniteList<TItem> = {
-  items: Array<TItem>;
-  totalCount: number;
-  isPending: boolean;
-  isError: boolean;
-  error: Error | null;
-  hasNextPage: boolean;
-  isFetchingNextPage: boolean;
-  fetchNextPage: () => void;
-};
-
-/**
- * Work out the offset the next page has to be requested at, based on the page that was received
- * last. Return `undefined` once the whole `totalCount` has been read, which tells the query that
- * there is nothing more to load.
- */
-export function getNextPageParam(lastPage: PagedResponse<unknown>) {
-  const { limit, offset, totalCount } = lastPage.pagination;
-  const next = offset + limit;
-
-  return next < totalCount ? next : undefined;
-}
 
 /**
  * Load a paged endpoint of the API one page at a time.
